requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.6
pymysql==1.1.1
pandas==2.2.3
requests==2.32.3
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
cryptography==46.0.5
